Address review comments
And move `Buffer` into its own module.
