+++ src/lib.rs
@@ -5,7 +5,6 @@
 #![allow(clippy::redundant_closure)]
 #![allow(clippy::type_complexity)]
 #![feature(async_await)]
-#![feature(async_closure)]
 #![cfg_attr(test, feature(specialization))]
 
 //! This crate provides a clean, ready to use client for [TiKV](https://github.com/tikv/tikv), a
